fix: use larger integer types for activity data

+++ backend/src/database/DB.go
@@ -325,8 +325,8 @@ func (db *DB) SeedTestData() {
 					UserID:     dbUsers[idx].ID,
 					CourseID:   courses[jdx].ID,
 					Type:       "interaction",
-					TotalTime:  uint(startTime + randTime),
-					TimeDelta:  uint(randTime),
+					TotalTime:  int64(startTime + randTime),
+					TimeDelta:  int64(randTime),
 					ExternalID: strconv.Itoa(rand.Intn(1000)),
 					CreatedAt:  time,
 				}
